Finish factory pather and fix everything
